// src/app/core/models/Product.ts
export interface Product {
    _id?: string;
    title: string;
    description: string;
    price: number;
    imageUrl: File;
    // category: string;
    // stockQuantity: number;
    // createdAt: Date;
    // updatedAt: Date;
    // isAvailable: boolean;
    // ratings?: number[];
    // averageRating?: number;
  }